BL_Backend/requirements.txt
Flask==2.0.1
SQLAlchemy==1.4
Flask-SQLAlchemy==2.5.1
flask-security-too
email_validator
bcrypt
flask_restful
redis
celery[redis]
Flask-Caching
flask_cors
weasyprint
pandas
reportlab
matplotlib
